requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comtypes==1.4.10
defusedxml==0.7.1
distro==1.9.0
EasyProcess==1.1
ecdsa==0.19.1
entrypoint2==1.1
fastapi==0.116.1
fpdf==1.7.2
frozenlist==1.7.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
inflection==0.5.1
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.0.0
multidict==6.6.3
numpy==2.2.6
openai==1.98.0
openapi==2.0.0
opencv-python==4.12.0.88
pefile==2024.8.26
pillow==11.3.0
propcache==0.3.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==1.0.1
python-dotenv==1.1.1
python-http-client==3.3.7
python-magic-bin==0.4.14
python-multipart==0.0.20
python-telegram-bot==20.0
pyttsx3==2.98
pytube==15.0.0
pytweening==1.2.0
pywin32==310
referencing==0.36.2
regex==2025.7.34
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.26.0
schedule==1.2.2
sendgrid==6.12.4
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.2
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.47.2
sympy==1.13.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.0
urllib3==2.4.0
uvicorn==0.35.0
Werkzeug==3.1.3
whisper==1.1.10
yarl==1.20.1
youtube-transcript-api==1.2.1
yt-dlp==2025.7.21
